adding routes for bathroom
